+++ app/controllers/users_controller.rb
@@ -6,8 +6,7 @@ class UsersController < ApplicationController
 
     def show 
         user = User.find(params[:id])
-        trees = Tree.all.select { |tree| tree.user_id == user.id}
-        render json: {id: user.id, name: user.name, trees: trees}
+        
     end 
 
     def create