+++ app/models/atmosphere.rb
@@ -1,4 +1,3 @@
 class Atmosphere < ApplicationRecord
     has_many :trees
-    has_many :users, through: :trees
 end