+++ app/models/tree.rb
@@ -1,4 +1,3 @@
 class Tree < ApplicationRecord
-    belongs_to :user
     belongs_to :atmosphere
 end