+++ app/models/user.rb
@@ -1,6 +1,5 @@
 class User < ApplicationRecord
-    has_many :trees 
-    has_many :atmospheres, through: :trees
+    has_many :atmospheres
 
     validates :name, uniqueness: true