+++ db/migrate/20200320144444_create_atmospheres.rb
@@ -1,10 +1,9 @@
 class CreateAtmospheres < ActiveRecord::Migration[6.0]
   def change
     create_table :atmospheres do |t|
-      t.integer :oxygen
-      t.integer :carbon_dioxide
-      t.integer :temperature
-
+      t.integer :oxygen, default: 30 
+      t.integer :carbon_dioxide, default: 25
+      t.integer :user_id
       t.timestamps
     end
   end